Tidy up of packet creation for S->C communication

+++ player/player.go
@@ -3,9 +3,10 @@ package player
 import "net"
 
 type Player struct {
+	// TODO: How should we manage all of these connections? -> The player probably doesn't need it directly
 	conn            *net.TCPConn
 	State           State
-	ProtocolVersion int
+	ProtocolVersion uint16
 	Username        string
 	Compression     CompressionState
 }
